Add filter to get most matched site or arbitrator
By design, arbitrator can not revoke tickets, and if the
commandline does not specify a site, it will try to find a "local site"
which has the same subnet with the machine that runs booth command.

If arbitrator and all sites are in the same subnet, booth will choose
the first one as "local site". Booth might think it is an arbitrator,
and will not execute revoke operations.

This patch add a filter to get most matched arbitrator/booth.