pacemaker: Use libpacemaker for setting and removing attributes.
libpacemaker provides a public API for managing tickets.  A first step
towards using this API is just the setting and removing code, which is
pretty straightforward.  We can ignore the XML result portion of the API
because in these cases it doesn't contain anything more useful than just
the return code.