pacemaker: Use libpacemaker to load ticket state.
The function itself is very simple.  The rest of the work is in
save_attributes, where it needs to understand both the libpacemaker XML
result as well as the result of calling crm_ticket directly.  It's still
pretty simple, though - just digging around in the XML to set a node
pointer correctly.

Fixes #136