build: Add support for building with mock.
The makefile target includes the name of the mock chroot you want to use
when building.  Most of the time for booth, you'd use it like so:

$ make mock-centos-stream+epel-9-x86_64

You may wish to change the release or arch, but due to the build
requirement on asciidoctor, it looks like epel always needs to be
included.